+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_commands.py
@@ -716,6 +716,7 @@ class ContainerappServiceBindingTests(ScenarioTest):
         redis_ca_name = 'redis'
         postgres_ca_name = 'postgres'
         kafka_ca_name = 'kafka'
+        mariadb_ca_name = 'mariadb'
 
         create_containerapp_env(self, env_name, resource_group)
 
@@ -728,6 +729,9 @@ class ContainerappServiceBindingTests(ScenarioTest):
         self.cmd('containerapp service kafka create -g {} -n {} --environment {}'.format(
             resource_group, kafka_ca_name, env_name))
 
+        self.cmd('containerapp service mariadb create -g {} -n {} --environment {}'.format(
+            resource_group, mariadb_ca_name, env_name))
+
         self.cmd('containerapp create -g {} -n {} --environment {} --image {} --bind postgres:postgres_binding redis'.format(
             resource_group, ca_name, env_name, image), checks=[
             JMESPathCheck('properties.template.serviceBinds[0].name', "postgres_binding"),
@@ -739,11 +743,12 @@ class ContainerappServiceBindingTests(ScenarioTest):
             JMESPathCheck('properties.template.serviceBinds[0].name', "redis"),
         ])
 
-        self.cmd('containerapp update -g {} -n {} --bind postgres:postgres_binding kafka'.format(
+        self.cmd('containerapp update -g {} -n {} --bind postgres:postgres_binding kafka mariadb'.format(
             resource_group, ca_name, image), checks=[
             JMESPathCheck('properties.template.serviceBinds[0].name', "redis"),
             JMESPathCheck('properties.template.serviceBinds[1].name', "postgres_binding"),
-            JMESPathCheck('properties.template.serviceBinds[2].name', "kafka")
+            JMESPathCheck('properties.template.serviceBinds[2].name', "kafka"),
+            JMESPathCheck('properties.template.serviceBinds[3].name', "mariadb")
         ])
 
         self.cmd('containerapp service postgres delete -g {} -n {} --yes'.format(
@@ -751,9 +756,12 @@ class ContainerappServiceBindingTests(ScenarioTest):
 
         self.cmd('containerapp service redis delete -g {} -n {} --yes'.format(
             resource_group, redis_ca_name, env_name))
-        
+
         self.cmd('containerapp service kafka delete -g {} -n {} --yes'.format(
             resource_group, kafka_ca_name, env_name))
+
+        self.cmd('containerapp service mariadb delete -g {} -n {} --yes'.format(
+            resource_group, mariadb_ca_name, env_name))
 
         self.cmd('containerapp service list -g {} --environment {}'.format(resource_group, env_name), checks=[
             JMESPathCheck('length(@)', 0),