+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/cluster/_deploy.py
@@ -13,7 +13,7 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "networkcloud cluster deploy",
-    is_experimental=True,
+    is_preview=True,
 )
 class Deploy(AAZCommand):
     """Deploy the cluster.
@@ -23,9 +23,9 @@ class Deploy(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2023-05-01-preview",
+        "version": "2023-07-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/clusters/{}/deploy", "2023-05-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/clusters/{}/deploy", "2023-07-01"],
         ]
     }
 
@@ -162,7 +162,7 @@ class Deploy(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-05-01-preview",
+                    "api-version", "2023-07-01",
                     required=True,
                 ),
             }