+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/storageappliance/_enable_remote_vendor_management.py
@@ -13,7 +13,7 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "networkcloud storageappliance enable-remote-vendor-management",
-    is_experimental=True,
+    is_preview=True,
 )
 class EnableRemoteVendorManagement(AAZCommand):
     """Enable remote vendor management of the provided storage appliance.
@@ -23,9 +23,9 @@ class EnableRemoteVendorManagement(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2023-05-01-preview",
+        "version": "2023-07-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/storageappliances/{}/enableremotevendormanagement", "2023-05-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.networkcloud/storageappliances/{}/enableremotevendormanagement", "2023-07-01"],
         ]
     }
 
@@ -152,7 +152,7 @@ class EnableRemoteVendorManagement(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-05-01-preview",
+                    "api-version", "2023-07-01",
                     required=True,
                 ),
             }