+++ src/communication/HISTORY.rst
@@ -3,9 +3,15 @@
 Release History
 ===============
 
+1.7.2b1
+++++++
+* Update Email to the latest 1.0.0 version
+
+
 1.7.2
 ++++++
 * Add additional test cases for ACS Rooms
+
 
 1.7.1b1
 ++++++